+++ scaleph-workflow/scaleph-workflow-api/src/main/java/cn/sliew/scaleph/workflow/engine/workflow/Dependent.java
@@ -18,5 +18,6 @@
 
 package cn.sliew.scaleph.workflow.engine.workflow;
 
-public interface Dependent {
+public interface Dependent extends Workflow {
+
 }
